# gen/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin-api/feedback/delete": {
            "post": {
                "description": "后台软删除用户反馈，仅标记 deleted_at，不物理删除数据。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "删除用户反馈",
                "parameters": [
                    {
                        "description": "删除反馈请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteFeedbackReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回 ok=true",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
        "/admin-api/feedback/detail": {
            "post": {
                "description": "后台根据反馈 ID 获取用户反馈详情，包含反馈内容、分类、状态、优先级、客户端环境、回复和处理信息。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "获取反馈详情",
                "parameters": [
                    {
                        "description": "反馈详情请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FeedbackDetailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回反馈详情",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "404": {
                        "description": "反馈不存在",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
        "/admin-api/feedback/list": {
            "post": {
                "description": "后台分页查询用户反馈，支持按 uid、分类、状态、优先级、关键词和创建时间范围筛选，默认按优先级和创建时间倒序返回。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "查询反馈列表",
                "parameters": [
                    {
                        "description": "反馈列表查询请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FeedbackListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回反馈分页列表",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
        "/admin-api/feedback/options": {
            "post": {
                "description": "获取反馈分类、处理状态和优先级枚举选项，用于后台筛选、表单下拉和前端展示。客户端也可复用该接口获取反馈分类。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "获取反馈枚举选项",
                "responses": {
                    "200": {
                        "description": "成功返回 categories/statuses/priorities",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
        "/admin-api/feedback/reply": {
            "post": {
                "description": "后台回复用户反馈。若请求未传 status，服务端默认将反馈状态置为已处理。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "回复用户反馈",
                "parameters": [
                    {
                        "description": "回复反馈请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReplyFeedbackReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回 ok=true",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误或状态非法",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
        "/admin-api/feedback/update-status": {
            "post": {
                "description": "后台更新用户反馈处理状态，可选更新优先级。状态：1待处理 2处理中 3已处理 4已关闭；优先级：1低 2普通 3高 4紧急。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "更新反馈状态",
                "parameters": [
                    {
                        "description": "更新反馈状态请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFeedbackStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回 ok=true",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误、状态非法或优先级非法",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
...
        "/v1/feedback/api/options": {
            "post": {
                "description": "获取反馈分类、处理状态和优先级枚举选项，用于后台筛选、表单下拉和前端展示。客户端也可复用该接口获取反馈分类。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AdminFeedback"
                ],
                "summary": "获取反馈枚举选项",
                "responses": {
                    "200": {
                        "description": "成功返回 categories/statuses/priorities",
                        "schema": {
                            "$ref": "#/definitions/response.DocResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.DocError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DeleteFeedbackReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "反馈ID",
                    "type": "integer"
                }
            }
        },
...
        "dto.FeedbackDetailReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "反馈ID",
                    "type": "integer"
                }
            }
        },
        "dto.FeedbackListReq": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "反馈分类",
                    "type": "integer"
                },
                "end_at": {
                    "description": "创建时间止，Unix秒",
                    "type": "integer"
                },
                "keyword": {
                    "description": "搜索标题/内容/联系方式",
                    "type": "string"
                },
                "page": {
                    "description": "页码，从1开始",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer"
                },
                "priority": {
                    "description": "优先级：1低 2普通 3高 4紧急",
                    "type": "integer"
                },
                "start_at": {
                    "description": "创建时间起，Unix秒",
                    "type": "integer"
                },
                "status": {
                    "description": "处理状态：1待处理 2处理中 3已处理 4已关闭",
                    "type": "integer"
                },
                "uid": {
                    "description": "业务用户UID",
                    "type": "integer"
                }
            }
        },
...
        "dto.ReplyFeedbackReq": {
            "type": "object",
            "required": [
                "id",
                "reply_content"
            ],
            "properties": {
                "id": {
                    "description": "反馈ID",
                    "type": "integer"
                },
                "reply_content": {
                    "description": "后台回复内容",
                    "type": "string",
                    "maxLength": 5000
                },
                "status": {
                    "description": "回复后状态，不传可由 service 默认置为已处理",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateFeedbackStatusReq": {
            "type": "object",
            "required": [
                "id",
                "status"
            ],
            "properties": {
                "id": {
                    "description": "反馈ID",
                    "type": "integer"
                },
                "priority": {
                    "description": "优先级：1低 2普通 3高 4紧急，不传则不修改",
                    "type": "integer"
                },
                "remark": {
                    "description": "处理备注，当前可作为扩展字段预留",
                    "type": "string",
                    "maxLength": 1000
                },
                "status": {
                    "description": "处理状态：1待处理 2处理中 3已处理 4已关闭",
                    "type": "integer"
                }
            }
        },
...
        "response.DocError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "msg_key": {
                    "type": "string"
                },
                "msg_params": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "rid": {
                    "type": "string"
                },
                "server_now": {
                    "type": "integer"
                }
            }
        },
        "response.DocResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                },
                "msg_key": {
                    "type": "string"
                },
                "msg_params": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "rid": {
                    "type": "string"
                },
                "server_now": {
                    "type": "integer"
                }
            }
        },
...
    }
}